move initialGameSettings to global settings
